ecommerce/services/product_service/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.2
psycopg2-binary==2.9.9
Pillow==10.4.0
PyJWT==2.9.0
gunicorn==22.0.0
